audio-summary-api/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
flask_mail==0.10.0
Werkzeug==2.2.2
openai-whisper @ git+https://github.com/openai/whisper.git
openai==0.27.8
torch==2.0.1
torchaudio==2.0.2
python-dotenv==1.0.0
numpy<2.0
gTTS==2.3.2
ffmpeg-python==0.2.0
